+++ test/mock/mock.test.ts
@@ -14,9 +14,7 @@ const IRL = {
 
 const Matrix = { tearDown: () => 1337, startup: () => 1 };
 
-const testSpyOn = (obj: Object, method: keyof typeof obj) => ({
-    returns: () => obj[method],
-});
+const testSpyOn = (obj: Object, method: keyof typeof obj) => obj[method];
 
 describe('Mocks', () => {
     beforeEach(() => {