+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2022-02-09
+### Added
+- [spy.getLatestCallArgument](//github.com/fdc-viktor-luft/spy4js#getLatestCallArgument)
+- [spy.getProps](//github.com/fdc-viktor-luft/spy4js#getProps)
+- [Spy.configure](//github.com/fdc-viktor-luft/spy4js#configure-static): New configuration option `useGenericReactMocks`.
+
 ## [3.2.0] - 2021-11-24
 ### Added
 - [spy.addSnapshotSerializer](//github.com/fdc-viktor-luft/spy4js#addSnapshotSerializer)